backend/requirements.txt
Flask==2.3.2
python-dotenv==0.21.0
anthropic==0.25.7
